+++ lightly_studio_view/src/lib/hooks/useOrderBy/useOrderBy.test.ts
@@ -1,151 +1,32 @@
 import { get, writable } from 'svelte/store';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { SortDirection } from '$lib/api/lightly_studio_local';
-import type { MetadataInfoView } from '$lib/api/lightly_studio_local';
-import type { EvaluationRunMetricsInfoView } from '$lib/api/lightly_studio_local/types.gen';
 import type { SortExpr } from '$lib/hooks/useImagesInfinite/useImagesInfinite';
+import type { SortField } from '$lib/hooks/useSortFields/useSortFields';
+import { IMAGE_SORT_FIELDS } from '$lib/hooks/useSortFields/useSortFields';
 import { useOrderBy } from './useOrderBy';
 
 const imageSortBy = writable<SortExpr[] | null>(null);
-const metadataInfo = writable<MetadataInfoView[]>([]);
-const metricsInfo = writable<{ data: EvaluationRunMetricsInfoView[] | null }>({ data: null });
+const allSortFields = writable<SortField[]>([...IMAGE_SORT_FIELDS]);
 const updateSortBy = vi.fn();
 
 vi.mock('$lib/hooks/useImageFilters/useImageFilters', () => ({
     useImageFilters: () => ({ imageSortBy, updateSortBy })
 }));
 
-vi.mock('$lib/hooks/useMetadataFilters/useMetadataFilters', () => ({
-    useMetadataFilters: () => ({ metadataInfo })
-}));
-
-vi.mock('$lib/hooks/useEvaluationSampleMetricsInfo/useEvaluationSampleMetricsInfo', () => ({
-    useEvaluationSampleMetricsInfo: () => metricsInfo
-}));
+vi.mock('$lib/hooks/useSortFields/useSortFields', async (importOriginal) => {
+    const original = await importOriginal<typeof import('$lib/hooks/useSortFields/useSortFields')>();
+    return {
+        ...original,
+        useSortFields: () => ({ allSortFields })
+    };
+});
 
 describe('useOrderBy', () => {
     beforeEach(() => {
         vi.clearAllMocks();
         imageSortBy.set(null);
-        metadataInfo.set([]);
-        metricsInfo.set({ data: null });
-    });
-
-    describe('allSortFields', () => {
-        it('contains the five base image sort fields', () => {
-            const { allSortFields } = useOrderBy({ datasetId: 'ds1' });
-            const fields = get(allSortFields);
-
-            expect(fields).toEqual(
-                expect.arrayContaining([
-                    expect.objectContaining({ source: 'image', value: 'file_name' }),
-                    expect.objectContaining({ source: 'image', value: 'file_path_abs' }),
-                    expect.objectContaining({ source: 'image', value: 'created_at' }),
-                    expect.objectContaining({ source: 'image', value: 'width' }),
-                    expect.objectContaining({ source: 'image', value: 'height' })
-                ])
-            );
-        });
-
-        it('includes metadata fields of supported types with metadata. prefix label', () => {
-            metadataInfo.set([
-                { name: 'score', type: 'float' },
-                { name: 'count', type: 'integer' },
-                { name: 'label', type: 'string' },
-                { name: 'active', type: 'boolean' }
-            ]);
-            const { allSortFields } = useOrderBy({ datasetId: 'ds1' });
-            const fields = get(allSortFields);
-
-            expect(fields).toEqual(
-                expect.arrayContaining([
-                    expect.objectContaining({
-                        source: 'metadata',
-                        value: 'score',
-                        label: 'metadata.score',
-                        is_numeric: true
-                    }),
-                    expect.objectContaining({
-                        source: 'metadata',
-                        value: 'count',
-                        label: 'metadata.count',
-                        is_numeric: true
-                    }),
-                    expect.objectContaining({
-                        source: 'metadata',
-                        value: 'label',
-                        label: 'metadata.label',
-                        is_numeric: false
-                    }),
-                    expect.objectContaining({
-                        source: 'metadata',
-                        value: 'active',
-                        label: 'metadata.active',
-                        is_numeric: false
-                    })
-                ])
-            );
-        });
-
-        it('excludes list and dict metadata fields', () => {
-            metadataInfo.set([
-                { name: 'tags', type: 'list' },
-                { name: 'nested', type: 'dict' },
-                { name: 'score', type: 'float' }
-            ]);
-            const { allSortFields } = useOrderBy({ datasetId: 'ds1' });
-            const fields = get(allSortFields);
-
-            expect(fields.map((f) => ('value' in f ? f.value : null))).not.toContain('tags');
-            expect(fields.map((f) => ('value' in f ? f.value : null))).not.toContain('nested');
-        });
-
-        it('includes evaluation metric fields with run_name_metric_name label', () => {
-            metricsInfo.set({
-                data: [
-                    {
-                        run_name: 'run1',
-                        metrics: [
-                            { metric_name: 'precision', min_value: 0, max_value: 1 },
-                            { metric_name: 'recall', min_value: 0, max_value: 1 }
-                        ]
-                    }
-                ]
-            });
-            const { allSortFields } = useOrderBy({ datasetId: 'ds1' });
-            const fields = get(allSortFields);
-
-            expect(fields).toEqual(
-                expect.arrayContaining([
-                    expect.objectContaining({
-                        source: 'evaluation_metric',
-                        evaluation_run_name: 'run1',
-                        metric_name: 'precision',
-                        label: 'run1_precision'
-                    }),
-                    expect.objectContaining({
-                        source: 'evaluation_metric',
-                        evaluation_run_name: 'run1',
-                        metric_name: 'recall',
-                        label: 'run1_recall'
-                    })
-                ])
-            );
-        });
-
-        it('updates reactively when metadataInfo changes', () => {
-            const { allSortFields } = useOrderBy({ datasetId: 'ds1' });
-
-            expect(
-                get(allSortFields).find((f) => 'value' in f && f.value === 'brightness')
-            ).toBeUndefined();
-
-            metadataInfo.set([{ name: 'brightness', type: 'float' }]);
-
-            expect(
-                get(allSortFields).find((f) => 'value' in f && f.value === 'brightness')
-            ).toBeDefined();
-        });
+        allSortFields.set([...IMAGE_SORT_FIELDS]);
     });
 
     describe('selectedDirection', () => {
@@ -188,7 +69,15 @@ describe('useOrderBy', () => {
         });
 
         it('returns the metadata label when a metadata field is active', () => {
-            metadataInfo.set([{ name: 'brightness', type: 'float' }]);
+            allSortFields.set([
+                ...IMAGE_SORT_FIELDS,
+                {
+                    source: 'metadata',
+                    value: 'brightness',
+                    label: 'metadata.brightness',
+                    is_numeric: true
+                }
+            ]);
             imageSortBy.set([
                 {
                     source: 'metadata',
@@ -344,7 +233,6 @@ describe('useOrderBy', () => {
         });
 
         it('sets is_numeric true for numeric metadata fields', () => {
-            metadataInfo.set([{ name: 'score', type: 'float' }]);
             const { handleFieldClick } = useOrderBy({ datasetId: 'ds1' });
 
             handleFieldClick({
